# Gamificator/Teaching-HEIGVD-AMT-2016-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Gamification platform
  description: |
    **API specification of the Gamification platform**

    You can try all HTTP operation described in this Swagger spec.
    
    Find source code of this API [here](https://github.com/sarrab/GamifyApp.git)
host: localhost:8090
basePath: /api
schemes:
  - http
  - https
paths:
...
  /pointScales/{pointScaleId}:
    get:
      description: Retourne le pointScale dont l'id est spécifié dans le path
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: pointScaleId
          description: PointScale's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Ok, PointScale details
          schema:
            $ref: '#/definitions/PointScaleDTO'
    put:
      description: Apporte des modifications au pointScale
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: pointScaleId
          description: PointScale's id
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Modification of the pointScale
          schema:
            $ref: '#/definitions/PointScaleDTO'
          required: false
      responses:
        200:
          description: OK, PointScale modified
    delete:
      description: Supprime le pointScale
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: pointScaleId
          description: PointScale's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK, PointScale deleted
...
  /rules:
    get:
      description: Retourne l'ensemble des règles de l'application
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
      responses:
        200:
          description:
              Application's rules
          schema:
            type: array
            items:
             $ref: '#/definitions/RuleDTO'
    post:
      description: crée une règle dans l'application
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: body
          name: body
          description: paramètres de la règles créée
          schema:
            $ref: '#/definitions/RuleDTO'
          required: true
      responses: 
        201:
          description: 
            OK règle créée
          schema:
            $ref: '#/definitions/RuleDTO'
  /rules/{ruleId}:
    get:
      description: Retourne la règle dont l'id est spécifié dans le path
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: ruleId
          description: rule's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Ok, Rule's details
          schema:
            $ref: '#/definitions/RuleDTO'
    put:
      description: Apporte des modifications à la règle
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: ruleId
          description: Rule's id
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Modification of the Rule
          schema:
            $ref: '#/definitions/RuleDTO'
          required: false
      responses:
        200:
          description: OK, Rule modified
    delete:
      description: Supprime une règle
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/TokenHeader'
        - in: path
          name: ruleId
          description: Rule's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK, Rule deleted
...
definitions:
...
  PointScaleDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        description: Badge name.
      description:
        type: string
        description: Description du badge
      nbrDePoints:
        type: integer
        format: int32
...
  RuleDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      event:
        type: string
        description: type d'évènement
      action:
        type: string
        description: type d'action
      pointScale:
        type: string
      points:
        type: integer
        format: int32
      badgeId:
        type: integer
        format: int64
      ActionId:
        type: integer
        format: int64
...
parameters:
  TokenHeader:
    description: token that identifies the app sending the request
    name: X-Gamification-Token
    in: header
    required: true
    type: string